Modified RPM section
I modified this section because we changed our sensor. We removed the proximity sensor and added a Hall sensor with magnets. I copied the reference code as is as we tested it on a small scale and it works but I think we should create a library for this section and tell the writer about it when we're done.